Fix the POSIX release legs: libc++ to_string ambiguity + non-portable test paths
Two of the three red legs on the v-tag release run were real (the third,
linux-arm64, was an apt mirror outage - rerun):
- macOS (clang/libc++): configStamp called std::to_string on the file-clock
  rep, which has no exact overload there (ambiguous). Explicit cast to
  long long.
- linux-x64: the P30 s7/s8 unit tests built their fixtures on literal
  "C:/proj" paths, which are RELATIVE paths on POSIX, so fs::absolute/
  fs::relative diverged and the closure-rule tests failed. Fixtures now
  build on temp_directory_path() so they are genuinely absolute everywhere.
Verified: Windows build + full unit suite green; WSL Ubuntu cmake/g++ build
green with all tests passing (the exact leg that was red).

+++ src/MintPlayer.Polyglot.Cli/src/pluginresolve.hpp
@@ -227,7 +227,8 @@ struct ResolveResult {
 inline std::string configStamp(const PgConfig& pc) {
     std::error_code ec;
     const auto t = std::filesystem::last_write_time(pc.dir / "pgconfig.json", ec);
-    return ec ? std::string("?") : std::to_string(t.time_since_epoch().count());
+    // Explicit cast: libc++'s file-clock rep has no exact to_string overload (ambiguous on macOS).
+    return ec ? std::string("?") : std::to_string(static_cast<long long>(t.time_since_epoch().count()));
 }
 
 // Resolve every dependency in a pgconfig (the `resolveConfiguredTargets` engine). `state` may be
